feat(database): add sample Postgres database blog post
This commit introduces a new blog post providing steps to quickly set up a sample Postgres database locally using Docker. The post includes an SQL script to initialize the database with tables and seed data, along with a simple Dockerfile and Makefile to build and run the database container. This will help developers get started with a pre-populated database for testing and development tasks.